Добавлена корректная обработка литров (0,45 /0,5 и тд) в weight - теперь Float вместо Int

+++ composeApp/src/commonMain/kotlin/com/mandarinkafe/mandarin/core/domain/models/Meal.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.Serializable
 data class Meal(
     val id: String,
     val name: String,
-    val weight: Int,
+    val weight: Float,
     val measureUnitType: MeasureUnitType,
     val price: Int,
     val mainImageUrl: String,
